apps/doctor-platform/doctor-api/requirements.txt
fastapi
uvicorn[standard]
boto3
python-dotenv
python-multipart
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
email_validator
requests
pytz 